stats_bot/requirements.txt
aiogram==3.11.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9

